02: update typography

+++ Typography/Typography.OpenFont/NetCore/Typography.OpenFont/CharacterMap.cs
@@ -51,12 +51,12 @@ namespace Typography.OpenFont
         }
         public ushort PlatformId { get; set; }
         public ushort EncodingId { get; set; }
-        public int CharacterToGlyphIndex(UInt32 character)
+        public ushort CharacterToGlyphIndex(UInt32 character)
         {
-            return (int)RawCharacterToGlyphIndex(character);
+            return RawCharacterToGlyphIndex(character);
         }
 
-        uint RawCharacterToGlyphIndex(UInt32 character)
+        ushort RawCharacterToGlyphIndex(UInt32 character)
         {
             // TODO: Fast segment lookup using bit operations?
             switch (this._cmapFormat)
@@ -72,7 +72,7 @@ namespace Typography.OpenFont
                                 if (_idRangeOffset[i] == 0)
                                 {
 
-                                    return (character + _idDelta[i]) % 65536;
+                                    return (ushort)((character + _idDelta[i]) % 65536);
                                 }
                                 else
                                 {
